+++ CHANGELOG.md
@@ -4,3 +4,5 @@ This changelog only contains the changes that are unreleased. For changes for in
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
 ## 3.4.0.1
+
+- Fix errors on OSX when files fail to download